Conflicts:
	we-web-parent/web-linyun-airline/src/main/webapp/WEB-INF/public/header.jsp
	we-web-parent/web-linyun-airline/src/main/webapp/public/dist/css/AdminLTE.css

+++ we-web-parent/web-linyun-airline/src/main/java/com/linyun/airline/forms/DictRelationUpdateForm.java
@@ -19,5 +19,5 @@ public class DictRelationUpdateForm extends ModForm implements Serializable {
 	private long targetId;
 
 	/**删除标识,0-未删除，1-已删除*/
-	private int dr;
+	private long dr;
 }
